Add function to parse human-readable part string
Part number and species are combined into a string to populate
multi-value lists, but the format can be broken down by regex to get the
part number and species again. The part type can be inferred from this
as well. This function returns all this information, though the caller
doesn't need to use all the fields.